refactor: move everything to bristlemouth_init
