+++ src/apps/bm_devkit/bmdk_common/app_main.cpp
@@ -18,7 +18,7 @@
 #include "app_pub_sub.h"
 #include "app_util.h"
 #include "bristlefin.h"
-#include "bristlemouth.h"
+#include "bristlemouth_client.h"
 #include "bsp.h"
 #include "cli.h"
 #include "reset_reason.h"
