+++ src/lib/bcmp/CMakeLists.txt
@@ -7,7 +7,8 @@ set(BCMP_FILES
     ${BCMP_DIR}/bcmp_cli.cpp
 
     # Core bristlemouth
-    ${BCMP_DIR}/bm/bristlemouth.cpp
+    ${BCMP_DIR}/bm/bm_printf.cpp
+    ${BCMP_DIR}/bm/bristlemouth_client.cpp
 
     PARENT_SCOPE)
